LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/script-src/manifest-v2-script-src-scheme-only-expected.txt

CONSOLE MESSAGE: Refused to load https://localhost:8443/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
ALERT: PASS: Successfully blocked script.

Test that scheme-only CSP for extensions blocks sources with that scheme.

PASS
